# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountRoleIn": {
        "description": "Тело `PUT /api/finance/accounts/{id}/role`.",
        "properties": {
          "role": {
            "description": "Одна из: checking, savings, unknown",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "AccountRoleIn",
        "type": "object"
      },
      "AccountsOut": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountOut"
            },
            "title": "Accounts",
            "type": "array"
          },
          "all_marked": {
            "description": "Ни одного счёта с ролью unknown: «Отложено» считается",
            "title": "All Marked",
            "type": "boolean"
          }
        },
        "required": [
          "accounts",
          "all_marked"
        ],
        "title": "AccountsOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Тело любого отказа API.\n\nИмя класса латиницей, в отличие от остального кода проекта, и это не\nвкусовщина: имена моделей Pydantic попадают в контракт именами схем\nOpenAPI, а из них генератор делает имена типов TypeScript. Кириллица\nтам формально законна и практически ломает автодополнение и генераторы.\nТо же правило действует в `api/schemas.py`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "JARVIS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/finance/accounts": {
      "get": {
        "description": "Счета с ролями. `all_marked=false` - «Отложено» пишет слова, а не сумму.",
        "operationId": "счета_api_finance_accounts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Нет действующего токена Cloudflare Access"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Токен принадлежит не owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Запрос не разобран"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "База данных или настройки недоступны"
          }
        },
        "summary": "Счета книжки и их роли",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/finance/accounts/{account_id}/role": {
      "put": {
        "description": "Разметка счетов owner (§15.5): по ней считается «Отложено».\n\nСчёт этой ручкой не заводится - он появляется импортом выписки. Поэтому\nадресуется id существующего счёта, а не пара «банк, имя»: разметить\nсчёт, которого нет, значит ошибиться в написании и получить двойника.",
        "operationId": "роль_счёта_api_finance_accounts__account_id__role_put",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRoleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Нет действующего токена Cloudflare Access"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Токен принадлежит не owner"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Операции, счёта или правила с таким id нет"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Запрос не разобран"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "База данных или настройки недоступны"
          }
        },
        "summary": "Роль счёта: расчётный или накопительный",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/finance/categories": {
      "get": {
        "description": "Набор категорий месяца - тот, из которого выбирают на карточке (§15.4).",
        "operationId": "категории_месяца_api_finance_categories_get",
        "parameters": [
          {
            "description": "Месяц в формате YYYY-MM. Не задан - текущий в зоне owner",
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Месяц в формате YYYY-MM. Не задан - текущий в зоне owner",
              "title": "Month"
            }
          },
          {
            "description": "Наборы всех месяцев сразу",
            "in": "query",
            "name": "all_months",
            "required": false,
            "schema": {
              "default": false,
              "description": "Наборы всех месяцев сразу",
              "title": "All Months",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Нет действующего токена Cloudflare Access"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Токен принадлежит не owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Запрос не разобран"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "База данных или настройки недоступны"
          }
        },
        "summary": "Категории месяца",
        "tags": [
          "finance"
        ]
      }
    },
...
    "/api/finance/overview": {
      "get": {
        "description": "Сальдо месяца с оговорками, без которых цифру читать нельзя (§15.5).\n\nПустой месяц - нули и накопленный итог предыдущих, а не отсутствие:\nкнижка, в которой этого месяца ещё нет, отвечает на вопрос «сколько\nпотрачено» честным нулём, а не 404.",
        "operationId": "обзор_месяца_api_finance_overview_get",
        "parameters": [
          {
            "description": "Месяц в формате YYYY-MM. Не задан - текущий в зоне owner",
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Месяц в формате YYYY-MM. Не задан - текущий в зоне owner",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Нет действующего токена Cloudflare Access"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Токен принадлежит не owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Запрос не разобран"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "База данных или настройки недоступны"
          }
        },
        "summary": "Обзор месяца: сальдо и статьи",
        "tags": [
          "finance"
        ]
      }
    },
...
  }
}